pactl, pacmd, cli-command: Added set-(sink|source)-latency-offset commands.
Added cli commands to both utilities to make use of the new
PA_COMMAND_SET_(SINK|SOURCE)_LATENCY_OFFSET protocol commands.

This enables applications to configure the latency of individual sources
and sinks via the command line interface.

The specific motivation is to enable network streaming applications which
interface with PulseAudio by capturing the output of a null sink, to have
PulseAudio correctly report the network latency (known by the application)
as part of the sink's total latency.