source, sink, core: Setting latency offsets of individual sources and sinks.
Some sinks and sources do not have ports associated with them, such as a null
sink. Nonetheless sources and sinks with no ports can be used for sound
input/output with an associated latency, for example by capturing output from
a null sink and directing over a network, which is how some implementations of
network audio streaming protocols interact with PulseAudio.

To keep graphics in sync with audio in these cases, the application/plugin must
be able to report the latency due to network transport/buffering to PulseAudio
and have it included in the total latency for that source or sink.

Previously, only the latency offset of a whole port could be set.
You could call pa_sink_set_latency_offset(), but the value you passed in would
be overwritten on change to a new port - that value was really just a cached
value of the port's latency offset. So those variables and functions have since
(in the previous commit) been renamed pa_sink_set_port_latency_offset() etc, to
distinguish them from the ones introduced in this commit.

This patch adds set_sink_latency_offset() and set_source_latency offset(),
and the corresponding members of the sink and source structs.

It also adds corresponding hooks PA_CORE_HOOK_SINK_LATENCY_OFFSET_CHANGED
and PA_CORE_HOOK_SOURCE_LATENCY_OFFSET_CHANGED.

The total latency reported by a source or sink now includes both the port
latency offset (if any) as well as the source or sink's own latency offset.